feat: added password changing via fzf
