+++ src/gns3util/api/delete_endpoints.py
@@ -1,4 +1,5 @@
-from . import GNS3APIClient
+# from . import GNS3APIClient
+from .client import GNS3APIClient
 
 
 class GNS3DeleteAPI(GNS3APIClient):
